+++ CHANGES.rst
@@ -5,6 +5,7 @@
   of the sample.
 - Update ``astropy_helpers`` to 1.0.2 and require astropy >= 1.0.2 to ensure
   pickleable Tables.
+- Internal data is a QTable instead of a dict.
 
 
 0.4 (2015-03-19)